fix the file name and content
