+++ examples/seeed-xiao-stm32c5/zephyr-can/zephyr-damiao-j4340P-2EC-v11-speed-fdcan1/platformio.ini
@@ -1,4 +1,4 @@
-; XIAO STM32C5 FDCAN2 minimal driver for DM-J4310-2EC V1.1
+; XIAO STM32C5 FDCAN1 driver for DM-J4340P-2EC V1.1 using an external transceiver
 
 [env:seeed-xiao-stm32c5]
 platform = https://github.com/Seeed-Studio/platform-seeedboards.git
