+++ examples/seeed-xiao-stm32c5/zephyr-can/zephyr-damiao-j4340P-2EC-v11-speed-fdcan1/zephyr/CMakeLists.txt
@@ -2,6 +2,6 @@
 
 cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(xiao_stm32c5_dm_j4310_2ec_speed)
+project(xiao_stm32c5_damiao_j4340p_2ec_v11_speed_fdcan1)
 
 target_sources(app PRIVATE ../src/main.c)
