+++ examples/seeed-xiao-stm32c5/zephyr-can/zephyr-damiao-j4340P-2EC-v11-speed-fdcan1/zephyr/prj.conf
@@ -1,3 +1,7 @@
+CONFIG_CAN=y
+CONFIG_CAN_INIT_PRIORITY=80
+CONFIG_CAN_DEFAULT_BITRATE=1000000
+
 CONFIG_GPIO=y
 CONFIG_POLL=y
 
@@ -11,8 +15,3 @@ CONFIG_USB_DEVICE_STACK_NEXT=y
 CONFIG_USBD_CDC_ACM_CLASS=y
 CONFIG_UART_LINE_CTRL=y
 CONFIG_UF2_DFU_RESET=y
-
-# ---- CAN bus (FDCAN2, classic CAN 2.0B @ 1 Mbps) ----
-CONFIG_CAN=y
-CONFIG_CAN_INIT_PRIORITY=80
-CONFIG_CAN_DEFAULT_BITRATE=1000000
