+++ examples/seeed-xiao-stm32c5/zephyr-can/zephyr-damiao-j4340P-2EC-v11-speed-fdcan2/platformio.ini
@@ -1,4 +1,4 @@
-; XIAO STM32C5 Zephyr Damiao DM-J4340P-2EC V1.1 24V speed sample, on-board FDCAN2
+; XIAO STM32C5 FDCAN2 driver for DM-J4340P-2EC V1.1 using the onboard transceiver
 
 [env:seeed-xiao-stm32c5]
 platform = https://github.com/Seeed-Studio/platform-seeedboards.git
