+++ examples/seeed-xiao-stm32c5/zephyr-can/zephyr-damiao-j4340P-2EC-v11-speed-fdcan2/src/main.c
@@ -1,11 +1,31 @@
 /*
  * SPDX-License-Identifier: Apache-2.0
  *
- * XIAO STM32C5 Damiao DM-J4340P-2EC V1.1 24V speed-mode sample.
+ * XIAO STM32C5 - DM-J4340P-2EC V1.1 speed-mode CAN driver.
+ *
+ * Single-file example that drives the motor over FDCAN2 (classic CAN 2.0B
+ * @ 1 Mbps) using the Damiao speed-mode protocol, and prints live feedback
+ * on the USB CDC ACM virtual COM port.
+ *
+ * Wiring (differential bus only — do NOT wire MCU CAN_TX/CAN_RX logic pins
+ * straight to the motor):
+ *   XIAO CANH -> motor CANH
+ *   XIAO CANL -> motor CANL
+ *   XIAO GND  -> motor GND (24V-)
+ *   24V+      -> motor VCC
+ *   (120 Ohm termination at each end of the bus)
+ *
+ * Protocol (Damiao speed mode, motor ID = 0x007):
+ *   Speed command : ID 0x200 + motor_id (0x207), D[0..3] = v_des float32 LE rad/s, DLC=4
+ *   Control cmd   : ID 0x200 + motor_id (0x207), D[0..6] = 0xFF, D[7] = cmd, DLC=8
+ *                    cmd: 0xFC = enable, 0xFD = disable, 0xFB = clear error
+ *   Parameter r/w : ID 0x7FF, D[0..1] = motor_id (LE), D[2] = op, D[3] = reg, ...
+ *                    op: 0x33 = read, 0x55 = write
+ *   Feedback      : standard frame, D[0] = motor_id | (fault<<4), D[1..7] =
+ *                    position / velocity / torque / temperatures (see decoder)
  */
 
 #include <stdint.h>
-#include <stdio.h>
 #include <string.h>
 
 #include <zephyr/device.h>
@@ -13,39 +33,57 @@
 #include <zephyr/drivers/can.h>
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/kernel.h>
+#include <zephyr/sys/printk.h>
 
-#define CANBUS_NODE DT_CHOSEN(zephyr_canbus)
+/* ---------------------------------------------------------------------- */
+/*  Motor / bus configuration                                             */
+/* ---------------------------------------------------------------------- */
 
-#ifndef DAMIAO_MOTOR_ID
-#define DAMIAO_MOTOR_ID 1U
-#endif
+/** Motor CAN ID (set by DIP switch / label on the motor). */
+#define DM_MOTOR_ID       0x007U
 
-#define DAMIAO_SPEED_MODE_OFFSET 0x200U
-#define DAMIAO_PARAM_FRAME_ID 0x7FFU
-#define DAMIAO_CONTROL_FRAME_ID (DAMIAO_SPEED_MODE_OFFSET + DAMIAO_MOTOR_ID)
-#define DAMIAO_CMD_ENABLE 0xFCU
-#define DAMIAO_CMD_DISABLE 0xFDU
-#define DAMIAO_CMD_CLEAR_ERROR 0xFBU
-#define DAMIAO_PARAM_READ 0x33U
-#define DAMIAO_PARAM_WRITE 0x55U
-#define DAMIAO_REG_ACC 0x04U
-#define DAMIAO_REG_DEC 0x05U
-#define DAMIAO_REG_MAX_SPD 0x06U
-#define DAMIAO_REG_CTRL_MODE 0x0AU
-#define DAMIAO_REG_VMAX 0x16U
-#define DAMIAO_REG_KP_ASR 0x19U
-#define DAMIAO_REG_KI_ASR 0x1AU
-#define DAMIAO_REG_DETA 0x1FU
-#define DAMIAO_REG_VBUS 0x3CU
-#define DAMIAO_CTRL_MODE_SPEED 3U
+/**
+ * Master (host) CAN ID. Stored in the motor as a parameter; the motor uses it
+ * to address parameter/feedback responses back to this host. It does not
+ * appear in the speed-mode command frame IDs.
+ */
+#define DM_MASTER_ID      0x017U
 
-#define CONTROL_PERIOD_MS 20
-#define GEAR_PERIOD_MS 5000
-#define FEEDBACK_PRINT_PERIOD_MS 500
+#define DM_CAN_BITRATE    1000000U
 
-static const struct device *const can_dev = DEVICE_DT_GET(CANBUS_NODE);
+/** Damiao speed-mode command base. Command/control ID = 0x200 + motor_id. */
+#define DM_SPEED_OFFSET   0x200U
+#define DM_CONTROL_ID     (DM_SPEED_OFFSET + DM_MOTOR_ID)  /* 0x207 */
+#define DM_PARAM_ID       0x7FFU
+
+/* Control commands (sent in D[7] of an 8-byte control frame). */
+#define DM_CMD_ENABLE        0xFCU
+#define DM_CMD_DISABLE       0xFDU
+#define DM_CMD_CLEAR_ERROR   0xFBU
+
+/* Parameter read/write opcodes and registers. */
+#define DM_PARAM_READ        0x33U
+#define DM_PARAM_WRITE       0x55U
+#define DM_REG_CTRL_MODE     0x0AU
+#define DM_CTRL_MODE_SPEED   3U
+
+/* Feedback velocity scaling upper bound (rad/s). Default VMAX for this motor. */
+#define DM_VMAX_RAD_S       30.0f
+
+/* ---------------------------------------------------------------------- */
+/*  Timing                                                               */
+/* ---------------------------------------------------------------------- */
+
+#define CONTROL_PERIOD_MS    20U    /* velocity command period (50 Hz)   */
+#define GEAR_PERIOD_MS       5000U  /* auto gear advance period          */
+#define FEEDBACK_PRINT_MS    500U   /* feedback print period             */
+
+/* ---------------------------------------------------------------------- */
+/*  Static state                                                         */
+/* ---------------------------------------------------------------------- */
+
+static const struct device *const can_dev = DEVICE_DT_GET(DT_CHOSEN(zephyr_canbus));
 static struct gpio_dt_spec led = GPIO_DT_SPEC_GET_OR(DT_ALIAS(led0), gpios, {0});
-static float feedback_vmax_rad_s = 30.0f;
 
 CAN_MSGQ_DEFINE(rx_msgq, 16);
 
@@ -55,121 +93,51 @@ struct speed_gear {
 	const char *label;
 };
 
+/* Demo speed profile. Adjust to taste; see the README for a fixed-speed note. */
 static const struct speed_gear gears[] = {
-	{0U, 0.0f, "stop"},
-	{1U, 3.0f, "low"},
-	{2U, 6.0f, "medium"},
+	{0U,  0.0f, "stop"},
+	{1U,  3.0f, "low"},
+	{2U,  6.0f, "medium"},
+	{3U, 10.0f, "high"},
 };
 
-static const uint8_t gear_sequence[] = {0U, 1U, 2U, 1U};
-static uint8_t sequence_pos;
+static const uint8_t gear_sequence[] = {0U, 1U, 2U, 3U, 2U, 1U};
+static uint8_t seq_pos;
 
-static const char *can_state_name(enum can_state state)
-{
-	switch (state) {
-	case CAN_STATE_ERROR_ACTIVE:
-		return "error-active";
-	case CAN_STATE_ERROR_WARNING:
-		return "error-warning";
-	case CAN_STATE_ERROR_PASSIVE:
-		return "error-passive";
-	case CAN_STATE_BUS_OFF:
-		return "bus-off";
-	case CAN_STATE_STOPPED:
-		return "stopped";
-	default:
-		return "unknown";
-	}
-}
+/* ---------------------------------------------------------------------- */
+/*  CAN helpers                                                          */
+/* ---------------------------------------------------------------------- */
 
-static void print_can_state(const char *prefix)
-{
-	enum can_state state;
-	struct can_bus_err_cnt err_cnt;
-	int ret;
-
-	ret = can_get_state(can_dev, &state, &err_cnt);
-	if (ret != 0) {
-		printf("%s CAN get_state failed: %d\n", prefix, ret);
-		return;
-	}
-
-	printf("%s CAN state=%s tx_err=%u rx_err=%u\n", prefix, can_state_name(state),
-	       err_cnt.tx_err_cnt, err_cnt.rx_err_cnt);
-}
-
-static void state_change_callback(const struct device *dev, enum can_state state,
-				  struct can_bus_err_cnt err_cnt, void *user_data)
+static void tx_callback(const struct device *dev, int error, void *user_data)
 {
 	ARG_UNUSED(dev);
 	ARG_UNUSED(user_data);
 
-	printf("CAN state change: state=%s tx_err=%u rx_err=%u\n", can_state_name(state),
-	       err_cnt.tx_err_cnt, err_cnt.rx_err_cnt);
-}
-
-static void tx_callback(const struct device *dev, int error, void *user_data)
-{
-	const char *label = user_data != NULL ? user_data : "unknown";
-
-	ARG_UNUSED(dev);
-
 	if (error != 0) {
-		printf("CAN TX callback %s error: %d\n", label, error);
-		print_can_state("tx callback");
+		printk("CAN TX error: %d\n", error);
 	}
 }
 
-static int damiao_can_send(const struct can_frame *frame, const char *label)
-{
-	int ret;
-
-	ret = can_send(can_dev, frame, K_NO_WAIT, tx_callback, (void *)label);
-	if (ret != 0) {
-		printf("CAN send %s immediate ret: %d id=0x%03x dlc=%u\n", label, ret,
-		       frame->id, frame->dlc);
-		print_can_state("send ret");
-	}
-
-	return ret;
-}
-
-static int damiao_send_cmd(uint8_t cmd)
+static int dm_send_cmd(uint8_t cmd)
 {
 	struct can_frame frame = {
-		.id = DAMIAO_CONTROL_FRAME_ID,
+		.id = DM_CONTROL_ID,
 		.dlc = 8,
 		.data = {0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, cmd},
 	};
 
-	return damiao_can_send(&frame, "cmd");
+	return can_send(can_dev, &frame, K_NO_WAIT, tx_callback, NULL);
 }
 
-static int damiao_read_param(uint8_t reg)
+static int dm_write_u32_param(uint8_t reg, uint32_t value)
 {
 	struct can_frame frame = {
-		.id = DAMIAO_PARAM_FRAME_ID,
-		.dlc = 4,
-		.data = {
-			DAMIAO_MOTOR_ID & 0xffU,
-			(DAMIAO_MOTOR_ID >> 8) & 0xffU,
-			DAMIAO_PARAM_READ,
-			reg,
-		},
-	};
-
-	return damiao_can_send(&frame, "read-param");
-}
-
-static int damiao_write_u32_param(uint8_t reg, uint32_t value)
-{
-	struct can_frame frame = {
-		.id = DAMIAO_PARAM_FRAME_ID,
+		.id = DM_PARAM_ID,
 		.dlc = 8,
 		.data = {
-			DAMIAO_MOTOR_ID & 0xffU,
-			(DAMIAO_MOTOR_ID >> 8) & 0xffU,
-			DAMIAO_PARAM_WRITE,
+			DM_MOTOR_ID & 0xffU,
+			(DM_MOTOR_ID >> 8) & 0xffU,
+			DM_PARAM_WRITE,
 			reg,
 			value & 0xffU,
 			(value >> 8) & 0xffU,
@@ -178,142 +146,57 @@ static int damiao_write_u32_param(uint8_t reg, uint32_t value)
 		},
 	};
 
-	return damiao_can_send(&frame, "write-param");
+	return can_send(can_dev, &frame, K_NO_WAIT, tx_callback, NULL);
 }
 
-static int damiao_send_speed(float speed_rad_s)
+static int dm_send_speed(float speed_rad_s)
 {
 	struct can_frame frame = {
-		.id = DAMIAO_CONTROL_FRAME_ID,
+		.id = DM_CONTROL_ID,
 		.dlc = 4,
 	};
 
 	memcpy(&frame.data[0], &speed_rad_s, sizeof(speed_rad_s));
-	return damiao_can_send(&frame, "speed");
+	return can_send(can_dev, &frame, K_NO_WAIT, tx_callback, NULL);
 }
 
-static void print_current_gear(void)
-{
-	const struct speed_gear *gear = &gears[gear_sequence[sequence_pos]];
-
-	printf("Gear %u (%s): %.2f rad/s\n", gear->gear, gear->label, (double)gear->speed_rad_s);
-}
-
-static void advance_gear(void)
-{
-	sequence_pos = (sequence_pos + 1U) % ARRAY_SIZE(gear_sequence);
-	print_current_gear();
-}
-
-static bool handle_param_response(const struct can_frame *frame)
-{
-	if (frame->dlc < 8U || frame->data[0] != (DAMIAO_MOTOR_ID & 0xffU) ||
-	    frame->data[1] != ((DAMIAO_MOTOR_ID >> 8) & 0xffU)) {
-		return false;
-	}
-
-	if (frame->data[2] == DAMIAO_PARAM_READ || frame->data[2] == DAMIAO_PARAM_WRITE) {
-		uint32_t value_u32 = (uint32_t)frame->data[4] | ((uint32_t)frame->data[5] << 8) |
-				     ((uint32_t)frame->data[6] << 16) |
-				     ((uint32_t)frame->data[7] << 24);
-		float value_float;
-
-		memcpy(&value_float, &value_u32, sizeof(value_float));
-
-		if (frame->data[3] == DAMIAO_REG_CTRL_MODE) {
-			printf("PARAM CTRL_MODE %s: %u\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write", value_u32);
-		} else if (frame->data[3] == DAMIAO_REG_ACC) {
-			printf("PARAM ACC %s: %.2f rad/s^2\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_DEC) {
-			printf("PARAM DEC %s: %.2f rad/s^2\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_MAX_SPD) {
-			printf("PARAM MAX_SPD %s: %.2f rad/s\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_VMAX) {
-			printf("PARAM VMAX %s: %.2f rad/s\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-			if (value_float > 0.0f && value_float < 1000.0f) {
-				feedback_vmax_rad_s = value_float;
-			}
-		} else if (frame->data[3] == DAMIAO_REG_VBUS) {
-			printf("PARAM VBus %s: %.2f V\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_KP_ASR) {
-			printf("PARAM KP_ASR %s: %.3f\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_KI_ASR) {
-			printf("PARAM KI_ASR %s: %.3f\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else if (frame->data[3] == DAMIAO_REG_DETA) {
-			printf("PARAM Deta %s: %.3f\n",
-			       frame->data[2] == DAMIAO_PARAM_READ ? "read" : "write",
-			       (double)value_float);
-		} else {
-			printf("PARAM reg=0x%02x cmd=0x%02x value_u32=%u value_float=%.3f\n",
-			       frame->data[3], frame->data[2], value_u32, (double)value_float);
-		}
-
-		return true;
-	}
-
-	return false;
-}
+/* ---------------------------------------------------------------------- */
+/*  Feedback decoding                                                    */
+/* ---------------------------------------------------------------------- */
 
 static void handle_feedback(void)
 {
 	static int64_t last_print_time;
-	static int64_t last_state_time;
 	struct can_frame frame;
 	int64_t now = k_uptime_get();
 
 	while (k_msgq_get(&rx_msgq, &frame, K_NO_WAIT) == 0) {
+		/* Speed-mode feedback is a standard frame with DLC >= 8. */
 		if ((frame.flags & CAN_FRAME_IDE) != 0U || frame.dlc < 8U) {
-			if ((frame.flags & CAN_FRAME_IDE) == 0U && frame.dlc >= 4U) {
-				printf("RX can_id=0x%03x dlc=%u data=%02x %02x %02x %02x\n",
-			       frame.id, frame.dlc, frame.data[0], frame.data[1],
-			       frame.data[2], frame.data[3]);
-			}
-			continue;
-		}
-
-		if (handle_param_response(&frame)) {
-			continue;
-		}
-
-		if (now - last_print_time < FEEDBACK_PRINT_PERIOD_MS) {
 			continue;
 		}
 
 		uint8_t motor_id = frame.data[0] & 0x0fU;
-		uint8_t err = frame.data[0] >> 4;
-		uint16_t pos_raw = ((uint16_t)frame.data[1] << 8) | frame.data[2];
-		uint16_t vel_raw = ((uint16_t)frame.data[3] << 4) | (frame.data[4] >> 4);
+		uint8_t fault    = frame.data[0] >> 4;
+		uint16_t pos_raw   = ((uint16_t)frame.data[1] << 8) | frame.data[2];
+		uint16_t vel_raw   = ((uint16_t)frame.data[3] << 4) | (frame.data[4] >> 4);
 		uint16_t torque_raw = ((uint16_t)(frame.data[4] & 0x0fU) << 8) | frame.data[5];
-		float vel_rad_s = (((float)vel_raw / 4095.0f) * 2.0f - 1.0f) * feedback_vmax_rad_s;
 
-		printf("FB can_id=0x%03x motor=%u err=0x%x pos=0x%04x vel=0x%03x %.2f rad/s tq=0x%03x mos=%u rotor=%u\n",
-		       frame.id, motor_id, err, pos_raw, vel_raw, (double)vel_rad_s, torque_raw,
-		       frame.data[6], frame.data[7]);
+		if (motor_id != (DM_MOTOR_ID & 0x0fU)) {
+			continue;
+		}
+
+		if (now - last_print_time < FEEDBACK_PRINT_MS) {
+			continue;
+		}
+
+		float vel_rad_s = (((float)vel_raw / 4095.0f) * 2.0f - 1.0f) * DM_VMAX_RAD_S;
+
+		printk("FB id=0x%03x motor=%u err=0x%x pos=0x%04x vel=0x%03x %.2f rad/s "
+		       "tq=0x%03x mos=%u rotor=%u\n",
+		       frame.id, motor_id, fault, pos_raw, vel_raw, (double)vel_rad_s,
+		       torque_raw, frame.data[6], frame.data[7]);
 		last_print_time = now;
-	}
-
-	if (last_state_time == 0) {
-		last_state_time = now;
-	}
-
-	if (now - last_state_time >= 1000) {
-		print_can_state("periodic");
-		last_state_time = now;
 	}
 }
 
@@ -327,83 +210,76 @@ static void drain_rx_for(int32_t duration_ms)
 	} while (k_uptime_get() < deadline);
 }
 
-static void read_param_with_drain(uint8_t reg)
-{
-	(void)damiao_read_param(reg);
-	drain_rx_for(80);
-}
+/* ---------------------------------------------------------------------- */
+/*  Main                                                                 */
+/* ---------------------------------------------------------------------- */
 
 int main(void)
 {
-	const struct can_filter feedback_filter = {
-		.id = 0,
-		.mask = 0,
-	};
 	int ret;
 
-	printf("XIAO STM32C5 Damiao DM-J4340P-2EC V1.1 24V speed sample\n");
-	printf("Automatic speed sequence every 5 seconds: 0->3->6->3->0 rad/s\n");
-	printf("Motor ID: %u, CAN bitrate: 1 Mbps\n", DAMIAO_MOTOR_ID);
+	printk("XIAO STM32C5 DM-J4340P-2EC V1.1 speed-mode driver\n");
+	printk("motor_id=0x%03x master_id=0x%03x bitrate=1 Mbps\n",
+	       DM_MOTOR_ID, DM_MASTER_ID);
 
 	if (!device_is_ready(can_dev)) {
-		printf("CAN device %s is not ready\n", can_dev->name);
+		printk("CAN device %s not ready\n", can_dev->name);
 		return 0;
 	}
-
-	can_set_state_change_callback(can_dev, state_change_callback, NULL);
-	print_can_state("initial");
 
 	if (led.port != NULL && gpio_is_ready_dt(&led)) {
 		(void)gpio_pin_configure_dt(&led, GPIO_OUTPUT_INACTIVE);
 	}
 
-	ret = can_set_bitrate(can_dev, 1000000U);
+	/* 1. Configure + start the FDCAN2 controller. */
+	ret = can_set_bitrate(can_dev, DM_CAN_BITRATE);
 	if (ret != 0) {
-		printf("Failed to set CAN bitrate: %d\n", ret);
+		printk("Failed to set CAN bitrate: %d\n", ret);
 		return 0;
 	}
 
-	ret = can_add_rx_filter_msgq(can_dev, &rx_msgq, &feedback_filter);
+	/* Catch-all RX filter: accept every frame, decode in handle_feedback(). */
+	const struct can_filter rx_all = { .id = 0, .mask = 0 };
+	ret = can_add_rx_filter_msgq(can_dev, &rx_msgq, &rx_all);
 	if (ret < 0) {
-		printf("Failed to add CAN RX filter: %d\n", ret);
+		printk("Failed to add CAN RX filter: %d\n", ret);
 		return 0;
 	}
 
 	ret = can_start(can_dev);
 	if (ret != 0) {
-		printf("Failed to start CAN controller: %d\n", ret);
+		printk("Failed to start CAN controller: %d\n", ret);
 		return 0;
 	}
-	print_can_state("after can_start");
 
-	(void)damiao_send_cmd(DAMIAO_CMD_CLEAR_ERROR);
+	/* 2. Motor startup sequence: clear error -> select speed mode -> enable. */
+	(void)dm_send_cmd(DM_CMD_CLEAR_ERROR);
 	drain_rx_for(100);
-	(void)damiao_write_u32_param(DAMIAO_REG_CTRL_MODE, DAMIAO_CTRL_MODE_SPEED);
+
+	(void)dm_write_u32_param(DM_REG_CTRL_MODE, DM_CTRL_MODE_SPEED);
 	drain_rx_for(100);
-	read_param_with_drain(DAMIAO_REG_CTRL_MODE);
-	read_param_with_drain(DAMIAO_REG_ACC);
-	read_param_with_drain(DAMIAO_REG_DEC);
-	read_param_with_drain(DAMIAO_REG_MAX_SPD);
-	read_param_with_drain(DAMIAO_REG_VMAX);
-	read_param_with_drain(DAMIAO_REG_KP_ASR);
-	read_param_with_drain(DAMIAO_REG_KI_ASR);
-	read_param_with_drain(DAMIAO_REG_DETA);
-	read_param_with_drain(DAMIAO_REG_VBUS);
-	(void)damiao_send_cmd(DAMIAO_CMD_ENABLE);
+
+	(void)dm_send_cmd(DM_CMD_ENABLE);
 	drain_rx_for(100);
-	print_current_gear();
+
+	printk("Motor enabled (speed mode). Cycling speed gears every %u s.\n",
+	       GEAR_PERIOD_MS / 1000U);
 
 	int64_t next_gear_time = k_uptime_get() + GEAR_PERIOD_MS;
 
 	while (true) {
-		const struct speed_gear *gear = &gears[gear_sequence[sequence_pos]];
+		const struct speed_gear *gear = &gears[gear_sequence[seq_pos]];
 
 		if (k_uptime_get() >= next_gear_time) {
-			advance_gear();
+			seq_pos = (seq_pos + 1U) % ARRAY_SIZE(gear_sequence);
+			printk("Gear %u (%s): %.2f rad/s\n",
+			       gears[gear_sequence[seq_pos]].gear,
+			       gears[gear_sequence[seq_pos]].label,
+			       (double)gears[gear_sequence[seq_pos]].speed_rad_s);
 			next_gear_time += GEAR_PERIOD_MS;
 		}
 
-		(void)damiao_send_speed(gear->speed_rad_s);
+		(void)dm_send_speed(gear->speed_rad_s);
 		handle_feedback();
 
 		if (led.port != NULL) {
